Display: Update to support 700 nit backlight
Changes our value from 3650 to 4095
This brings our max brightness to a usable value in sunlight.

Note that we should not stay above 3300 or 550 nit for extended periods of time. 3300 was measured to be within the safe zone of our display spec. Values above 3300 will increase the display wear significantly and lead to burn in. We should only use this zone when absolutely necessary ie direct sunlight. We should test the burn in on an old unit we do not care about, as of now I believe HBM is only safe for at a maximum of 1 hour.

Change-Id: I8d3f07bdf0420ab89b724be9addbe61506787229
Signed-off-by: Nick Franco 👽 <nfranco@osomprivacy.com>