cape: Add osom,tui-argon memory heap for Argon2 in TUI
Argon2 in TUI requires exactly 256 MiB of memory to run. To avoid
permanently carving out that memory for the TVM, add a CMA heap that TUI
can tap on the fly via the hypervisor to gain access to the 256 MiB needed
to run Argon2.

Although Argon2 only needs 256 MiB, make the CMA heap larger than that by
128 MiB (for a total of 384 MiB) to add a guard band in case the CMA
allocator encounters pages that it cannot migrate out from the heap's
region when TUI needs that memory. The guard band likely doesn't need to be
this large, but failing to allocate the 256 MiB for Argon2 when TUI needs
it is catastrophic, so the guard band is intentionally paranoid.

There is no alignment requirement for this memory, so the alignment page
order is configured to 0 to increase the odds of successfully allocating
the needed memory when some pages in the region cannot be migrated.

Change-Id: Ifa0270a3e8f17ba9713a3daa3f102f655084f692
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>