waipio-vm: Increase CPU allocation to TVM to four cores
To run Argon2 in TUI with as much computational power as can be afforded,
increase the CPU allocation to TVM from two cores to four cores. This adds
CPU3 and CPU7 to the TVM.

This is the maximum number of cores that can be allocated to TVM, as
evidenced by hyp_core_ctl's limit of `num_possible_cpus()/2`. Although this
limit could be bypassed in hyp_core_ctl in the PVM, bypassing it doesn't
make it possible to allocate more than four CPUs to TVM, possibly due to a
limitation in Gunyah.

Although it would be preferred to give CPU4 to TVM instead of CPU3, it's
not possible to give up all of the gold CPUs to TVM because at least one
gold core must remain to support 32-bit EL0. The gold CPUs are the only
ones with 32-bit EL0 support, hence why CPU4 must remain with the PVM. CPU4
could still technically be used, but it'd have to be shared simultaneously
between the PVM and TVM when TUI runs, which is suboptimal for performance
because neither VM's scheduler is aware of the other VM using the CPU.

Change-Id: I614296e82bbfed15e22b83642b1dc618b7b0fe7e
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>